/Refactor : item 파트의 클래스 분할
