+++ src/main/java/com/trioshop/controller/item/CartController.java
@@ -4,6 +4,7 @@ import com.trioshop.SessionConst;
 import com.trioshop.model.dto.item.CartEntity;
 import com.trioshop.model.dto.item.ItemInfoByCart;
 import com.trioshop.model.dto.user.UserInfoBySession;
+import com.trioshop.service.item.CartService;
 import com.trioshop.service.item.ItemService;
 import com.trioshop.utils.CategoryList;
 import lombok.NoArgsConstructor;
@@ -21,13 +22,13 @@ import java.util.List;
 @RequiredArgsConstructor
 public class CartController {
 
-    private final ItemService itemService;
+    private final CartService cartService;
 
     @GetMapping("/cart") // 카트 페이지로
     public String cartPage(@SessionAttribute(SessionConst.LOGIN_MEMBER) UserInfoBySession userInfoBySession,
                            Model model) {
 
-        List<ItemInfoByCart> cartItems = itemService.cartItemList(userInfoBySession.getUserCode());
+        List<ItemInfoByCart> cartItems = cartService.cartItemList(userInfoBySession.getUserCode());
         model.addAttribute("cartItems", cartItems);
         return "user/itemInfo/cart";
     }
@@ -37,7 +38,7 @@ public class CartController {
                               @RequestParam("cartItemQty") long cartItemQty,
                               @SessionAttribute(SessionConst.LOGIN_MEMBER) UserInfoBySession userInfoBySession,
                               Model model) {
-        itemService.addCartItem(new CartEntity(userInfoBySession.getUserCode(), itemCode, cartItemQty));
+        cartService.addCartItem(new CartEntity(userInfoBySession.getUserCode(), itemCode, cartItemQty));
         //return "user/itemInfo/itemList";
         return "redirect:itemList";
     }
@@ -46,7 +47,7 @@ public class CartController {
                                @RequestParam(value = "quantities", required = false) List<Long> quantities,
                                @SessionAttribute(SessionConst.LOGIN_MEMBER) UserInfoBySession userInfoBySession,
                                Model model) {
-        itemService.addCartItems(userInfoBySession.getUserCode(), itemCodes, quantities);
+        cartService.addCartItems(userInfoBySession.getUserCode(), itemCodes, quantities);
         return "redirect:itemList";
     }
     @PostMapping("/cart/remove") // 카트 품목 제거
@@ -54,7 +55,7 @@ public class CartController {
                                   @SessionAttribute(SessionConst.LOGIN_MEMBER) UserInfoBySession userInfoBySession,
                                   Model model ){
         //userCode 와 itemCode 만으로 이루어진 생성자 호출
-        itemService.deleteCartItem(new CartEntity(userInfoBySession.getUserCode(), itemCode));
+        cartService.deleteCartItem(new CartEntity(userInfoBySession.getUserCode(), itemCode));
         return "redirect:/cart";
     }
 }
