+++ src/main/java/com/trioshop/repository/dao/item/ItemInfoDao.java
@@ -26,7 +26,6 @@ public class ItemInfoDao {
     public List<ItemInfoByUser> searchItems(ItemCondition itemCondition) {
         return itemMapper.searchItems(itemCondition);
     }
-
     public List<ItemInfoByUser> itemInfoByCodes(List<Long> itemCodes) {
         return itemMapper.itemInfoByCodes(itemCodes);
     }
@@ -35,49 +34,7 @@ public class ItemInfoDao {
         return itemMapper.itemInfoByCode(itemCode);
     }
 
-    public List<ItemInfoByCart> cartItemList(long userCode) {
-        return itemMapper.cartItemList(userCode);
-    }
-
-    public List<OrderListModel> orderList(long userCode) {
-        return itemMapper.orderList(userCode);
-    }
-
-    public void insertOrders(OrdersEntity ordersEntity) {
-        itemMapper.insertOrders(ordersEntity);
-    }
-
-    public void insertOrderItems(OrderItemEntity orderItemEntity) {
-        itemMapper.insertOrderItems(orderItemEntity);
-    }
-
-    public void updateStockQty(ItemCodeAndQty itemCodeAndQty) {
-        itemMapper.updateStockQty(itemCodeAndQty);
-    }
-
-    public void deleteItemsFromCart(long userCode, List<Long> itemCodeList) {
-        itemMapper.deleteItemsFromCart(userCode, itemCodeList);
-    }
-
-    public void insertCartItem(CartEntity cartEntity) {
-        itemMapper.insertCartItem(cartEntity);
-    }
-
-    public void deleteCartItem(CartEntity cartEntity) {
-        itemMapper.deleteCartItem(cartEntity);
-    }
-
-    public int selectCartItem(CartEntity cartEntity) {
-        return itemMapper.selectCartItem(cartEntity);
-    }
-
-    public void updateCartItem(CartEntity cartEntity) {
-        itemMapper.updateCartItem(cartEntity);
-    }
     public List<ItemDetailSearch> itemDetailNameSearch(String itemName) {
         return itemMapper.itemDetailNameSearch(itemName);
-    }
-    public UserAddressInfo selectUserAddressInfo (long userCode) {
-        return  itemMapper.selectUserAddressInfo(userCode);
     }
 }
